Added auth

+++ src/main/java/org/example/components/empolyee/EmployeeService.java
@@ -1,36 +1,68 @@
 package org.example.components.empolyee;
 
+import org.example.auth.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.Optional;
 
 @Service
 public class EmployeeService {
     private final EmployeeRepository employeeRepository;
     private final MyEmployeeRepository myEmployeeRepository;
+    private final UserService userService;
     @Autowired
-    public EmployeeService(EmployeeRepository employeeRepository, MyEmployeeRepository myEmployeeRepository) {
+    public EmployeeService(EmployeeRepository employeeRepository, MyEmployeeRepository myEmployeeRepository,
+                           UserService userService) {
         this.employeeRepository = employeeRepository;
         this.myEmployeeRepository = myEmployeeRepository;
+        this.userService = userService;
     }
 
-    public Iterable<Employee> findAll(){
-        return employeeRepository.findAll();
+    public ResponseEntity<String> findAll(HttpServletRequest request) {
+        if (!userService.checkAdmin(request)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
+        }
+        Iterable<Employee> employees = employeeRepository.findAll();
+        return ResponseEntity.ok(employees.toString());
     }
-    public void save(Employee employee){
+    public ResponseEntity<String> save(Employee employee, HttpServletRequest request){
+        if (!userService.checkAdmin(request)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
+        }
         employeeRepository.save(employee);
+        return ResponseEntity.ok("Saved");
     }
 
-    public void deleteById(Long id){
+    public ResponseEntity<String>  deleteById(Long id, HttpServletRequest request){
+        if (!userService.checkAdmin(request)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
+        }
         employeeRepository.deleteById(id);
+        return ResponseEntity.ok("Deleted");
     }
 
-    public Iterable<Employee> filter(String name, String email, String phone, Double salary, String position){
-        return myEmployeeRepository.filter(name, email, phone, salary, position);
+    public ResponseEntity<String> filter(String name, String email, String phone, Double salary, String position,
+                                         HttpServletRequest request){
+        if (!userService.checkAdmin(request)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
+        }
+        Iterable<Employee> employees = myEmployeeRepository.filter(name, email, phone, salary, position);
+        return ResponseEntity.ok(employees.toString());
     }
 
-    public Optional<Employee> findById(Long id){
-        return employeeRepository.findById(id);
+    public ResponseEntity<String> findById(Long id, HttpServletRequest request){
+        if (!userService.checkAdmin(request)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
+        }
+        Optional<Employee> employee = employeeRepository.findById(id);
+        if (!employee.isPresent()){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
+        }else{
+            return ResponseEntity.ok(employee.toString());
+        }
     }
 }
